requirements.txt
discord.py==2.3.1
apscheduler==3.10.1